Setup project for heroku hosting and change seeder location
