+++ Program.cs
@@ -3,9 +3,9 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.DependencyInjection;
-using MvcMovie.Models;
-using MvcActor.Models;
-using MvcSeries.Models;
+// using MvcMovie.Models;
+// using MvcActor.Models;
+// using MvcSeries.Models;
 using ElectronNET.API;
 
 namespace IMDB
@@ -14,26 +14,7 @@ namespace IMDB
     {
         public static void Main(string[] args)
         {
-             var host = CreateHostBuilder(args).Build();
-
-            using (var scope = host.Services.CreateScope())
-            {
-                var services = scope.ServiceProvider;
-
-                try
-                {
-                  SeedMovies.Initialize(services);
-                  SeedSeries.Initialize(services);
-                  SeedActors.Initialize(services);
-                }
-                catch (Exception ex)
-                {
-                    var logger = services.GetRequiredService<ILogger<Program>>();
-                    logger.LogError(ex, "An error occurred seeding the DB.");
-                }
-            }
-
-            host.Run();
+            CreateHostBuilder(args).Build().Run();
         }
 
         public static string HostPort => Environment.GetEnvironmentVariable("PORT")??"5000";
