+++ appsettings.json
@@ -6,10 +6,5 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "MvcMovieContext": "Data Source=MvcMovie.db",
-    "MvcSeriesContext": "Data Source=MvcSeries.db",
-    "MvcActorContext": "Data Source=MvcActor.db"
-  }
+  "AllowedHosts": "*"
 }
